refactor(cli): extract manifest_edit module
add_secret_to_manifest (and its validate_add_secret_name helper) lived
in the cli module behind the cli feature, which also pulls in
clap_complete, clap_complete_nushell, and is_executable for shell
completions. A caller that only wants to add a secret declaration to
a manifest has no way to take that capability without the completion
tooling too.

Moved both functions into a new manifest_edit module behind its own
manifest-edit feature (which cli continues to enable, so cli's
behavior is unchanged). manifest-edit depends on toml_edit alone.
