+++ secretspec/src/cli/mod.rs
@@ -567,28 +567,9 @@ fn generate_toml_with_comments(config: &Config) -> crate::Result<String> {
     Ok(doc.to_string())
 }
 
-/// Rejects names that cannot occupy a flattened secret key in [`Profile`].
-fn validate_add_secret_name(name: &str) -> Result<()> {
-    if !crate::config::is_valid_identifier(name) {
-        return Err(miette!(
-            "Invalid secret name '{}': must be a valid identifier (alphanumeric and underscores, not starting with a number)",
-            name
-        ));
-    }
-    // `Profile` reserves this key for its defaults table before flattening all
-    // remaining keys into secret declarations. Without an explicit check, the
-    // edit would be valid TOML but would not actually declare a secret.
-    if name == "defaults" {
-        return Err(miette!(
-            "Secret name 'defaults' is reserved for profile defaults"
-        ));
-    }
-    Ok(())
-}
-
 /// Ensures `add` will create a new effective declaration in an existing profile.
 fn validate_add_target(app: &Secrets, profile: &str, name: &str) -> Result<()> {
-    validate_add_secret_name(name)?;
+    crate::manifest_edit::validate_add_secret_name(name)?;
 
     if !app.config().profiles.contains_key(profile) {
         let mut available: Vec<&str> = app.config().profiles.keys().map(String::as_str).collect();
@@ -608,57 +589,6 @@ fn validate_add_target(app: &Secrets, profile: &str, name: &str) -> Result<()> {
     }
 
     Ok(())
-}
-
-/// Adds one secret to a manifest document without re-serializing the rest.
-///
-/// `toml_edit` retains the user's comments, whitespace, ordering, and any syntax
-/// that is not represented by [`Config`]. The caller validates the selected
-/// profile against the fully loaded configuration first; this helper creates a
-/// local profile table when that profile currently comes only from `extends`.
-fn add_secret_to_manifest(
-    source: &str,
-    profile: &str,
-    name: &str,
-    description: &str,
-) -> Result<String> {
-    use toml_edit::{DocumentMut, InlineTable, Item, Table, Value};
-
-    validate_add_secret_name(name)?;
-    if description.trim().is_empty() {
-        return Err(miette!("Secret description cannot be empty"));
-    }
-
-    let mut doc = source
-        .parse::<DocumentMut>()
-        .into_diagnostic()
-        .wrap_err("Failed to parse secretspec.toml for editing")?;
-    let profiles = doc
-        .get_mut("profiles")
-        .and_then(Item::as_table_like_mut)
-        .ok_or_else(|| miette!("secretspec.toml does not contain a [profiles] table"))?;
-
-    if !profiles.contains_key(profile) {
-        profiles.insert(profile, Item::Table(Table::new()));
-    }
-    let profile_table = profiles
-        .get_mut(profile)
-        .and_then(Item::as_table_like_mut)
-        .ok_or_else(|| miette!("Profile '{}' is not a TOML table", profile))?;
-
-    if profile_table.contains_key(name) {
-        return Err(miette!(
-            "Secret '{}' is already declared in profile '{}'",
-            name,
-            profile
-        ));
-    }
-
-    let mut secret = InlineTable::new();
-    secret.insert("description", Value::from(description));
-    profile_table.insert(name, toml_edit::value(secret));
-
-    Ok(doc.to_string())
 }
 
 /// Replaces an existing manifest only after its complete replacement has been
@@ -1063,7 +993,12 @@ pub fn main() -> Result<()> {
             let source = fs::read_to_string(&manifest_path)
                 .into_diagnostic()
                 .wrap_err_with(|| format!("Failed to read {}", manifest_path.display()))?;
-            let updated = add_secret_to_manifest(&source, &profile, &name, description)?;
+            let updated = crate::manifest_edit::add_secret_to_manifest(
+                &source,
+                &profile,
+                &name,
+                description,
+            )?;
             write_manifest_atomically(&manifest_path, &updated)?;
 
             println!(
@@ -2287,57 +2222,6 @@ mod tests {
     }
 
     #[test]
-    fn add_secret_to_manifest_preserves_comments_and_other_tables() {
-        let source = r#"# Project documentation
-[project]
-name = "demo"
-revision = "1.0"
-
-[profiles.default]
-# Keep this explanation attached to the existing secret.
-DATABASE_URL = { description = "Database connection string" }
-
-[providers]
-local = "dotenv://.env"
-"#;
-
-        let updated =
-            add_secret_to_manifest(source, "default", "API_KEY", "API access token").unwrap();
-
-        assert!(updated.contains("# Project documentation"));
-        assert!(updated.contains("# Keep this explanation attached to the existing secret."));
-        assert!(updated.contains("local = \"dotenv://.env\""));
-        assert!(updated.contains("API_KEY = { description = \"API access token\" }"));
-
-        let config: Config = toml::from_str(&updated).expect("edited manifest must parse");
-        config.validate().expect("edited manifest must validate");
-        assert_eq!(
-            config.profiles["default"].secrets["API_KEY"]
-                .description
-                .as_deref(),
-            Some("API access token")
-        );
-    }
-
-    #[test]
-    fn add_secret_to_manifest_can_overlay_an_inherited_profile() {
-        let source = r#"[project]
-name = "demo"
-revision = "1.0"
-extends = ["../shared"]
-
-[profiles.default]
-LOCAL = { description = "Local secret" }
-"#;
-
-        let updated =
-            add_secret_to_manifest(source, "production", "API_KEY", "API access token").unwrap();
-
-        assert!(updated.contains("[profiles.production]"));
-        assert!(updated.contains("API_KEY = { description = \"API access token\" }"));
-    }
-
-    #[test]
     fn add_target_rejects_inherited_secrets_and_unknown_profiles() {
         let config: Config = toml::from_str(
             r#"[project]
@@ -2364,33 +2248,6 @@ API_KEY = { description = "API access token" }
         assert!(unknown.contains("Available profiles: default, development"));
 
         validate_add_target(&app, "development", "NEW_KEY").unwrap();
-    }
-
-    #[test]
-    fn add_secret_to_manifest_rejects_invalid_or_duplicate_declarations() {
-        let source = r#"[profiles.default]
-API_KEY = { description = "Existing" }
-"#;
-
-        let invalid = add_secret_to_manifest(source, "default", "1BAD", "Description")
-            .unwrap_err()
-            .to_string();
-        assert!(invalid.contains("Invalid secret name"));
-
-        let reserved = add_secret_to_manifest(source, "default", "defaults", "Description")
-            .unwrap_err()
-            .to_string();
-        assert!(reserved.contains("reserved for profile defaults"));
-
-        let duplicate = add_secret_to_manifest(source, "default", "API_KEY", "Description")
-            .unwrap_err()
-            .to_string();
-        assert!(duplicate.contains("already declared"));
-
-        let empty = add_secret_to_manifest(source, "default", "NEW_KEY", "   ")
-            .unwrap_err()
-            .to_string();
-        assert!(empty.contains("description cannot be empty"));
     }
 
     #[test]
